requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
jinja2>=3.1.0

# Ingestion
pymupdf>=1.25.0
pymupdf4llm>=0.0.17
python-frontmatter>=1.1.0

# Embeddings & Vector Store
sentence-transformers>=3.4.0
chromadb>=0.6.0

# LLM Providers
openai>=1.60.0
google-genai>=1.5.0

# UI
streamlit>=1.42.0
requests>=2.32.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
httpx>=0.28.0
